requirements.txt
fastapi
uvicorn
python-dotenv
pytest
ruff
httpx
